GI-9 Validate email is from justice.gov.uk domain
When signing up the users email address should only be valid if the
email address is from the justice.gov.uk domain. This includes the
digital.justice.gov.uk domain and other subdomains.

To verify this validation in the real world, email confirmation will
need to be turned on in Devise.  This will be done in a separate PR.

Existing tests have been updated to use valid email addresses.