GI-56 Add email functionality to app
This adds Gov.uk notify email messaging.  The GOVUK_NOTIFY_API_KEY
environment variable should be set before starting the server for the
emails to work.

API Keys should be stored securely, there is a test key (doesn't send
email) and a whitelist key (sends to team members on notify).

When tests are running the perform_deliveries = false prevents
deliveries from being made externally.

The email_spec.rb is largely redundant when this setting is enabled but
I will leave it here for the moment as it is useful to test the notify
service until we have something else in the app that sends an email, by
setting the perform_deliveries to true (or commenting out)

The notify mailer is currently configured to send a simple template,
this will be updated in future PRs to add in personalisation and
integration with devise.