GI-57 Add log out link
Log out link added to top of all pages if the current user is logged in.
This should be moved later to a better position eg nav bar when we have
one.

Also tidied up the urls for the users/sign_in etc to just /sign_in as
the user part at the start is a little untidy.